Fix Votes controller. Disable AD auth. Add user as manual input

+++ Awards/Controllers/VotesController.cs
@@ -75,28 +75,46 @@ namespace Awards.Controllers
         }
 
         // POST: api/Votes
-        [ResponseType(typeof(Vote))]
-        public async Task<IHttpActionResult> PostVote(Vote vote)
+        [ResponseType(typeof(void))]
+        public async Task<IHttpActionResult> PostVote(string user, SetVoteDTO vote)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
+            var currentUser = ""; //Get from Claims
+            var nominee = await db.Nominees.FindAsync(vote.NomineeID);
+            if(nominee.Votes.Any(o => o.Voter == currentUser))
+            {
+                var response = new HttpResponseMessage(HttpStatusCode.Forbidden);
+                response.ReasonPhrase = "You have already voted for a nominee in this category";
+                return ResponseMessage(response);
+            }
 
-            db.Votes.Add(vote);
+            db.Votes.Add(new Vote
+            {
+                NomineeID = vote.NomineeID,
+                Voter = currentUser
+            });
             await db.SaveChangesAsync();
 
-            return CreatedAtRoute("DefaultApi", new { id = vote.ID }, vote);
+            return Ok();
         }
 
         // DELETE: api/Votes/5
         [ResponseType(typeof(Vote))]
-        public async Task<IHttpActionResult> DeleteVote(int id)
+        public async Task<IHttpActionResult> DeleteVote(string user, int id)
         {
             Vote vote = await db.Votes.FindAsync(id);
             if (vote == null)
             {
                 return NotFound();
+            }
+            if (vote.Voter != user)
+            {
+                var response = new HttpResponseMessage(HttpStatusCode.Forbidden);
+                response.ReasonPhrase = "You cannot delete other peoples votes.";
+                return ResponseMessage(response);
             }
 
             db.Votes.Remove(vote);
